Add a Candidate struct with a .seed function
Candidates are represented as bitmaps, that
represent which permutations have been seen, as
well as the length of unique symbols in the tail
of the string.

We can get away with N-2 bits for the tail of the
string because there are N-2 possible outcomes.
The tail of the string will always be a maximum of
N-1 symbols and we won’t expand candidates with
duplicate symbols next to each other. Here are the
counts of bits we’d set for N=5:

123451 -> 3451 -> 2
123452 -> 3452 -> 2
123453 -> 453  -> 1
123454 -> 54   -> 0

Given these Bitmaps will be relatively large and
we plan on storing a lot of them, we use Roaring
Bitmaps to efficiently store them in memory.