Add a Symmetry#mapping function
This function determines which mapping should be
used (if there is more than one option) to
canonicalise the candidate’s bitmap. It chooses
the mapping that would result in a more leftmost
bit being set in the resulting bitmap.

For example:

- 100000 is preferred over 010000
- 110000 is preferred over 101000
- 000110 is preferred over 000011

You can think of this as choosing the bitmap with
largest integer representation:

- 32 is preferred over 16
- 48 is preferred over 40
- 6 is preferred over 3