Implement Eq and Ord for Candidate
This is so we can create a BTreeSet of candidates.
We could use a HashSet, but the hashing algorithm
is likely to be slower than this.

The ordering uses the same idea as the canonical
ordering for symmetries. The candidate that has a
bit set further to the left is deemed ‘less than’
the other. This doesn’t really matter, as long as
it’s consistent.