Implement the A* search algorithm
The heuristic is currently just some boilerplate
so the #shortest_path function can be written as
intended.

The #shortest_path function actually only returns
the length of the shortest path, rather than a
series of candidates. This can be upgraded later,
but for now this saves memory.

The goal can be any candidate that has the given
number of bits. This actually means the goal is
part of an ‘abstract state space’, which we’ll use
more later to improve the heuristic.

If we do reach a goal, we need to leave the open
and closed sets in a consistent state so we can
resume again with the next goal number of bits.
Normally, A* would just return immediately, but we
need to make sure we finish expanding the node.