Allow tests to override the number of symbols
I want to test the search algorithm against four
symbols as this should be relatively fast, but the
number of symbols is a static variable.

This change adds an Interface struct which allows
the static variable to be set dynamically and uses
a compiler feature to override it explicitly for
tests that want to use four symbols.

This is a bit clunky as it requires opting out of
four_symbols at the top of test files, but it’ll
do for now.