Reindex buckets when the heuristics change
When the heuristics change, this affects the
h_cost for each candidate, which in turn affects
the f_cost.

The f_cost is used to determine which order
candidates should be expanded in the A* search, so
it’s important we update it, otherwise we wouldn’t
use the new information provided by the improved
heuristic.