Increase lower bounds for candidates with fewer than starting bits
If a candidate immediately goes in the wrong
direction it needs to add a number of bits back
again, on top of those it also needs to reach the
next goal.

Given we know what the shortest distance is to add
some number of bits, I think we can increase their
lower bounds based on these - though I’m not 100%
sure whether this is correct.

I don’t think it will actually affect anything
because I don’t think these candidates will ever
have the lowest f_cost, so the A* search will
never expand them.