Fix Heuristic to make it admissible
Previously, the lower bound on the remaining
distance to the goal was wrong, in that it was
sometimes returning a value greater than the true
distance to the goal. This causes A* to find a
path that isn’t necessarily the shortest.

There were two problems:

1) Basing the distance on the number of bits
   didn’t take into account additional distance
   that could be travelled when starting from a
   non-full set of counter bits.

2) The current search depth of the candidate
   wasn’t being taken into account and the lower
   bound was based on the incorrect assumption
   that the candidate had the minimum depth for
   some number of bits.

The algorithm now finds the known shortest
distances for the 4 and 5 symbol cases which
indicates it is working correctly.